More import fixes
