+++ ml-agents-envs/mlagents/envs/base_unity_environment.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Dict
 
-from mlagents.envs import AllBrainInfo, BrainParameters
+from mlagents.envs.brain import AllBrainInfo, BrainParameters
 
 
 class BaseUnityEnvironment(ABC):
