+++ ml-agents-envs/mlagents/envs/env_manager.py
@@ -1,6 +1,8 @@
 from abc import ABC, abstractmethod
 from typing import List, Dict, NamedTuple, Optional
-from mlagents.envs import AllBrainInfo, BrainParameters, Policy, ActionInfo
+from mlagents.envs.brain import AllBrainInfo, BrainParameters
+from mlagents.envs.policy import Policy
+from mlagents.envs.action_info import ActionInfo
 
 
 class EnvironmentStep(NamedTuple):
