--- ml-agents/mlagents/trainers/__init__.py
@@ -1,20 +1,0 @@
-from .buffer import *
-from .curriculum import *
-from .meta_curriculum import *
-from .models import *
-from .trainer_metrics import *
-from .trainer import *
-from .tf_policy import *
-from .trainer_controller import *
-from .bc.models import *
-from .bc.offline_trainer import *
-from .bc.online_trainer import *
-from .bc.policy import *
-from .ppo.models import *
-from .ppo.trainer import *
-from .ppo.policy import *
-from .sac.models import *
-from .sac.trainer import *
-from .sac.policy import *
-from .exception import *
-from .demo_loader import *
