--- ml-agents/mlagents/trainers/components/bc/__init__.py
@@ -1,1 +1,0 @@
-from .module import BCModule
