+++ ml-agents/mlagents/trainers/components/bc/model.py
@@ -1,5 +1,4 @@
 import tensorflow as tf
-import numpy as np
 from mlagents.trainers.models import LearningModel
