+++ ml-agents/mlagents/trainers/components/reward_signals/curiosity/signal.py
@@ -3,7 +3,6 @@ import numpy as np
 import tensorflow as tf
 from mlagents.envs.brain import BrainInfo
 
-from mlagents.trainers.buffer import Buffer
 from mlagents.trainers.components.reward_signals import RewardSignal, RewardSignalResult
 from mlagents.trainers.components.reward_signals.curiosity.model import CuriosityModel
 from mlagents.trainers.tf_policy import TFPolicy
