--- ml-agents/mlagents/trainers/components/reward_signals/extrinsic/__init__.py
@@ -1,1 +1,0 @@
-from .signal import ExtrinsicRewardSignal
