+++ ml-agents/mlagents/trainers/components/reward_signals/extrinsic/signal.py
@@ -2,7 +2,6 @@ from typing import Any, Dict, List
 import numpy as np
 from mlagents.envs.brain import BrainInfo
 
-from mlagents.trainers.buffer import Buffer
 from mlagents.trainers.components.reward_signals import RewardSignal, RewardSignalResult
 from mlagents.trainers.tf_policy import TFPolicy
 from mlagents.trainers.models import LearningModel
