--- ml-agents/mlagents/trainers/components/reward_signals/gail/__init__.py
@@ -1,1 +1,0 @@
-from .signal import GAILRewardSignal
