+++ ml-agents/mlagents/trainers/components/reward_signals/reward_signal_factory.py
@@ -2,7 +2,7 @@ import logging
 from typing import Any, Dict, Type
 
 from mlagents.trainers.trainer import UnityTrainerException
-from mlagents.trainers.components.reward_signals.reward_signal import RewardSignal
+from mlagents.trainers.components.reward_signals import RewardSignal
 from mlagents.trainers.components.reward_signals.extrinsic.signal import (
     ExtrinsicRewardSignal,
 )
