+++ ml-agents/mlagents/trainers/learn.py
@@ -14,9 +14,9 @@ from typing import Any, Callable, Dict, Optional
 
 from mlagents.trainers.trainer_controller import TrainerController
 from mlagents.trainers.exception import TrainerError
-from mlagents.trainers import MetaCurriculumError, MetaCurriculum
+from mlagents.trainers.meta_curriculum import MetaCurriculumError, MetaCurriculum
 from mlagents.trainers.trainer_util import initialize_trainers
-from mlagents.envs import UnityEnvironment
+from mlagents.envs.environment import UnityEnvironment
 from mlagents.envs.sampler_class import SamplerManager
 from mlagents.envs.exception import UnityEnvironmentException, SamplerException
 from mlagents.envs.base_unity_environment import BaseUnityEnvironment
