+++ ml-agents/mlagents/trainers/tests/test_curriculum.py
@@ -2,7 +2,7 @@ import pytest
 from unittest.mock import patch, mock_open
 
 from mlagents.trainers.exception import CurriculumError
-from mlagents.trainers import Curriculum
+from mlagents.trainers.curriculum import Curriculum
 
 
 dummy_curriculum_json_str = """
