+++ ml-agents/mlagents/trainers/tests/test_learn.py
@@ -1,6 +1,7 @@
 import pytest
 from unittest.mock import MagicMock, patch
-from mlagents.trainers import learn, TrainerController
+from mlagents.trainers import learn
+from mlagents.trainers.trainer_controller import TrainerController
 
 
 @pytest.fixture
